Split src/client/admin.ts into per-concern modules under src/client/admin/
The single admin bundle entrypoint was ~660 lines mixing 15 unrelated
behaviors. Each behavior now lives in its own module with an explicit
initXxx() function; admin.ts becomes a thin wiring file that calls each
initializer. Modules self-guard for missing DOM elements so they remain
safe to run on every admin page.
