+++ src/client/admin/closes-at-autofill.ts
@@ -1,0 +1,16 @@
+/// <reference lib="dom" />
+/** Auto-populate closes_at from event date when closes_at is empty. */
+export const initClosesAtAutofill = (): void => {
+  const dateInput =
+    document.querySelector<HTMLInputElement>('input[name="date"]');
+  const closesAtInput = document.querySelector<HTMLInputElement>(
+    'input[name="closes_at"]',
+  );
+  if (!dateInput || !closesAtInput) return;
+
+  dateInput.addEventListener("change", () => {
+    if (dateInput.value && !closesAtInput.value) {
+      closesAtInput.value = dateInput.value;
+    }
+  });
+};
