+++ src/client/admin/nav-select.ts
@@ -1,0 +1,11 @@
+/// <reference lib="dom" />
+/** Navigate to selected option value on change. */
+export const initNavSelect = (): void => {
+  for (const el of document.querySelectorAll<HTMLSelectElement>(
+    "[data-nav-select]",
+  )) {
+    el.addEventListener("change", () => {
+      location.href = el.value;
+    });
+  }
+};
