+++ src/client/admin/select-on-click.ts
@@ -1,0 +1,9 @@
+/// <reference lib="dom" />
+/** Auto-select input contents when clicked. */
+export const initSelectOnClick = (): void => {
+  for (const el of document.querySelectorAll<HTMLInputElement>(
+    "[data-select-on-click]",
+  )) {
+    el.addEventListener("click", () => el.select());
+  }
+};
